% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_linestrings.R
\name{break_one_line}
\alias{break_one_line}
\title{Load an sf object with l feature consisting of a multilinestring geometry. This function is used in break_to_smallest()}
\usage{
break_one_line(line, geom = "geometry", crs = 27700)
}
\arguments{
\item{line}{An sf object or sf dataframe with 1 multilinestring row/feature}

\item{geom}{Field containing the active geometry}

\item{crs}{Default is 27700}
}
\value{
Returns the same sf object with the multilinestring replaced with smaller linestrings
}
\description{
This function loads a sf object with 1 feature, the active geometry "multilinestring"
and breaks each of these lines to many points that make up the lines and stitches them back up
to linestrings.
}
