% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_linestrings.R
\name{break_to_smallest}
\alias{break_to_smallest}
\title{Load an sf object with linestring geometry.}
\usage{
break_to_smallest(lines, messaging = 5)
}
\arguments{
\item{lines}{An sf object or sf dataframe}

\item{messaging}{Progress indicator displaying in Percentages.}
}
\value{
Returns the same sf object with the addition of line geometries gotten from the broken larger linestrings
}
\description{
This function loads an sf object with the active geometry "linestring" and breaks each of these larger lines to many lines.
}
