% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertCoordinates.R
\name{east_north_to_long_lat}
\alias{east_north_to_long_lat}
\title{Load a list consiting easting and northing}
\usage{
east_north_to_long_lat(easting, northing, crs = 27700)
}
\arguments{
\item{easting}{Easting is the eastward-measured distance (or the x-coordinate)}

\item{northing}{Northing is the northward-measured distance (or the y-coordinate)}

\item{crs}{Default is 27700 for United Kingdom}
}
\value{
Returns a dataframe with easting and northing converted to longitude and latitude
}
\description{
This function loads a list or dataframe containing easting and northing and outputs longitude and latitude using the provided crs.
}
