% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_linestrings.R
\name{fix_xymax}
\alias{fix_xymax}
\title{Fix the xmax and/or ymax when the xmax and xmin are the same and/or the ymax and ymin are the same.}
\usage{
fix_xymax(x)
}
\arguments{
\item{x}{An sf object with the active geometry type "linestring" which is a pair of points and a matrix of 4 values representing 2 points.}
}
\value{
Returns the same sf object with the fixed xmax and/or ymax.
}
\description{
This function loads an sf object with the active geometry type "linestring" and adds 0.1 to at least one of the xmax or ymax if the xmax and xmin are the same or the ymax and ymin are the same.
}
