% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bearing_per_row.R
\name{get_bearing_per_row}
\alias{get_bearing_per_row}
\title{Calculate the number of times to split/divide each linestring.}
\usage{
get_bearing_per_row(df = df)
}
\arguments{
\item{x}{An sf object or sf dataframe consisting of a column denoting the length of geometry}

\item{max_length}{Maximum length for each linestring. Can also be a floating value.}

\item{length_col}{The field with the length variable}

\item{messaging}{Progress indicator displaying in Percentages}
}
\value{
Returns the same R object with the addition of "splits_in_meters" columns
}
\description{
This function loads an R object with the active geometry "linestring" and breaks each of these multilines to many lines.
}
