% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Load a dataframe}
\usage{
get_elevation(
  x,
  AssetID = "AssetID",
  Longitude = "Longitude",
  Latitude = "Latitude",
  unit = "meters",
  src = "epqs"
)
}
\arguments{
\item{x}{An R object or a dataframe}

\item{AssetID}{An asset base containing the AssetID, Longitude and Latitude}

\item{Longitude}{numerical string(s) consisting geographic coordinate specific to the east-west
position of a point on the Earth's surface}

\item{Latitude}{numerical string(s) consisting geographic coordinate specific to the north-south
position of a point on the Earth's surface}

\item{unit}{elevation unit in "meters" or "feet"}

\item{src}{this selects which API to use. "epqs"is best for US only, "aws" for large number of points (>500)}
}
\value{
Returns the same dataframe with the addition of the "elevation" column
}
\description{
This function loads a dataframe. It accepts an R object with unique asset
identifier (default set as "AssetID"), longitude and latitude to calculate elevation.
}
