% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_feature_column.R
\name{get_nearest_feature_column}
\alias{get_nearest_feature_column}
\title{Load 2 sf objects, x and y and the fields you want to add from y to x}
\usage{
get_nearest_feature_column(x, y, col = "column_1")
}
\arguments{
\item{x}{An sf object containing the features for which you want find the nearest fields in y}

\item{y}{An sf object containing fields to be assigned to x}

\item{col}{The column(s) to add from y to x. Default is "column_1"}
}
\value{
Returns the same sf object in x with the addition of the column(s) from y
}
\description{
Accepts two sf objects x and y and appends the columns of the nearest feature in y to
every feature in x. I use "fields" and "columns" interchangeably.
}
