% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_to_one_distance.R
\name{get_one_to_one_distance}
\alias{get_one_to_one_distance}
\title{Load 1 sf object which contains the spatial feature of both the 'x geometry' and the 'y geometry' columns}
\usage{
get_one_to_one_distance(
  x,
  x_geom = "geometry.x",
  y_geom = "geometry.y",
  splits = 50,
  centroid = TRUE
)
}
\arguments{
\item{x}{The sf object with both the x geometry and the y geometry}

\item{x_geom}{The geometry of the object on the lhs}

\item{y_geom}{The geometry of the object on the rhs}

\item{splits}{The number of lists so as to speed up the calculation}

\item{centroid}{Default is TRUE. If TRUE, convert x_geom and y_geom to centroids}
}
\value{
Returns the sf object with additional column, geom_distance_m in meters
}
\description{
This function will calculate the distance between the 2 spatial features in each row
}
