% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertCoordinates.R
\name{long_lat_to_east_north}
\alias{long_lat_to_east_north}
\title{Load a list consiting long (longitude) and lat (latitude)}
\usage{
long_lat_to_east_north(long, lat, crs = 27700)
}
\arguments{
\item{long}{numerical string(s) consisting geographic coordinate specific to the east-west
position of a point on the Earth's surface}

\item{lat}{numerical string(s) consisting geographic coordinate specific to the north-south
position of a point on the Earth's surface}

\item{crs}{Default is 27700 for United Kingdom}
}
\value{
Returns a dataframe with longitude and latitude converted to easting and northing
}
\description{
This function loads a list or dataframe containing easting and northing and outputs longitude and latitude using the provided crs.
}
