% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_linestrings.R
\name{split_my_linestrings}
\alias{split_my_linestrings}
\title{Split the geometry using the splits_number column}
\usage{
split_my_linestrings(
  x,
  length_col = "length_m",
  splits_number = "splits_in_meters",
  uid = "pipe_id"
)
}
\arguments{
\item{x}{An sf object or sf dataframe with a length column (length_col), splits_number, and a unique idenifier.}

\item{length_col}{The field with the continuous "length" variable which denotes the length of the geometry.}

\item{splits_number}{The field with the length(in meters) to split the geometry with.}

\item{uid}{This column represents the unique identifier to identify the geometries which have or have not been split.}
}
\value{
Returns a dataframe with the addition of "new_length" column
}
\description{
This function loads an sf object with the active geometry "linestring" and breaks each of these large lines to many lines.
}
